Moves all components and utils in HistoryPage to its own files
