+++ react/src/features/history/components/HistoryTitleComponent.tsx
@@ -1,0 +1,5 @@
+export default function HistoryTitleComponent({ title }: { title: string }) {
+  return (
+    <h2 className="text-3xl font-semibold text-black font-ibm-plex-sans">{title}</h2>
+  );
+}
