+++ react/src/features/history/typedefs/DummyType.ts
@@ -1,0 +1,4 @@
+export type DummyType = {
+  name: string,
+  timestamp: number,
+};
